+++ cloudpathlib/azure/azblobpath.py
@@ -1,7 +1,7 @@
 import os
 from pathlib import Path
 from tempfile import TemporaryDirectory
-from typing import TYPE_CHECKING
+from typing import Any, Optional, TYPE_CHECKING
 
 from cloudpathlib.exceptions import CloudPathIsADirectoryError
 
@@ -39,10 +39,10 @@ class AzureBlobPath(CloudPath):
     def drive(self) -> str:
         return self.container
 
-    def mkdir(self, parents=False, exist_ok=False):
+    def mkdir(self, parents=False, exist_ok=False, mode: Optional[Any] = None):
         self.client._mkdir(self, parents=parents, exist_ok=exist_ok)
 
-    def touch(self, exist_ok: bool = True):
+    def touch(self, exist_ok: bool = True, mode: Optional[Any] = None):
         if self.exists():
             if not exist_ok:
                 raise FileExistsError(f"File exists: {self}")
@@ -56,7 +56,7 @@ class AzureBlobPath(CloudPath):
 
             tf.cleanup()
 
-    def stat(self):
+    def stat(self, follow_symlinks=True):
         try:
             meta = self.client._get_metadata(self)
         except ResourceNotFoundError: